# docs/architecture/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "本机管理系统当前 HTTP 契约",
    "version": "2026-08-08-working-tree",
    "description": "从当前Flask路由提取。x-legacy-exception表示请求/响应尚未统一；字段继续以源码与测试为准。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "默认本机Flask"
    }
  ],
...
  "paths": {
...
    "/api/proxy-pool/status": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_proxy_pool_status_get",
        "summary": "GET /api/proxy-pool/status",
        "x-source": "gateway/app.py:8076",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/accounts": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_accounts_route_get",
        "summary": "GET /api/accounts",
        "x-source": "gateway/app.py:8096",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/accounts/save": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_save_account_route_post",
        "summary": "POST /api/accounts/save",
        "x-source": "gateway/app.py:8100",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/account/next": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_next_account_get",
        "summary": "GET /api/account/next",
        "x-source": "gateway/app.py:8116",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/account/update": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_update_account_route_post",
        "summary": "POST /api/account/update",
        "x-source": "gateway/app.py:8125",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/accounts/discover": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_discover_accounts_route_post",
        "summary": "POST /api/accounts/discover",
        "x-source": "gateway/app.py:8148",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/accounts/import": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_import_accounts_route_post",
        "summary": "POST /api/accounts/import",
        "x-source": "gateway/app.py:8161",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/accounts/proxy": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_assign_account_proxy_route_post",
        "summary": "POST /api/accounts/proxy",
        "x-source": "gateway/app.py:8185",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/browser-v2/comment-campaigns": {
      "get": {
        "tags": [
          "Comment Campaign"
        ],
        "operationId": "comment_campaign_list_campaigns_get",
        "summary": "GET /api/browser-v2/comment-campaigns",
        "x-source": "comment_campaign/blueprint.py:155",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "tags": [
          "Comment Campaign"
        ],
        "operationId": "comment_campaign_create_campaign_post",
        "summary": "POST /api/browser-v2/comment-campaigns",
        "x-source": "comment_campaign/blueprint.py:169",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "已创建的Campaign",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CommentCampaignEnvelope"}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CampaignCreate"}
            }
          }
        }
      }
    },
...
    "/api/selector-probe/gates": {
      "get": {
        "tags": [
          "Selector Probe"
        ],
        "operationId": "selector_probe_gates_route_get",
        "summary": "GET /api/selector-probe/gates",
        "x-source": "selector_probe/blueprint.py:2543",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "认证模式Flask Session；local-direct模式由loopback/Host guard保护。"
      }
    },
    "schemas": {
...
      "CommentCampaignRead": {
        "type": "object",
        "description": "Campaign公开读取投影；identity_generation由全量账户预检生成，客户端不得写入。",
        "properties": {
          "id": {"type": "string", "readOnly": true},
          "name": {"type": "string", "readOnly": true},
          "mode": {"type": "string", "readOnly": true},
          "target_source": {"type": "string", "readOnly": true},
          "target_reference": {"type": "string", "readOnly": true},
          "video_id": {"type": "string", "readOnly": true},
          "canonical_url": {"type": "string", "readOnly": true},
          "template_id": {"type": "string", "readOnly": true},
          "template_revision": {"type": "integer", "minimum": 1, "readOnly": true},
          "profile_refs": {"type": "array", "items": {"type": "string"}, "readOnly": true},
          "batch_size": {"type": "integer", "readOnly": true},
          "allocation_seed": {"type": "string", "readOnly": true},
          "start_mode": {"type": "string", "readOnly": true},
          "scheduled_at": {"type": ["string", "null"], "readOnly": true},
          "status": {"type": "string", "readOnly": true},
          "pause_reason": {"type": ["string", "null"], "readOnly": true},
          "revision": {"type": "integer", "minimum": 1, "readOnly": true},
          "prepare_generation": {"type": "integer", "minimum": 0, "readOnly": true},
          "identity_generation": {"type": "integer", "minimum": 0, "readOnly": true}
          ,"template_snapshot": {"type": "object", "readOnly": true},
          "profile_snapshot": {"type": "array", "readOnly": true},
          "content_snapshot": {"type": ["array", "object"], "readOnly": true},
          "locked_at": {"type": ["string", "null"], "readOnly": true},
          "created_at": {"type": "string", "readOnly": true},
          "updated_at": {"type": "string", "readOnly": true}
        },
        "required": ["id", "name", "mode", "target_source", "target_reference", "video_id", "canonical_url", "template_id", "template_revision", "profile_refs", "batch_size", "allocation_seed", "start_mode", "scheduled_at", "status", "pause_reason", "revision", "prepare_generation", "identity_generation", "template_snapshot", "profile_snapshot", "content_snapshot", "locked_at", "created_at", "updated_at"],
        "additionalProperties": false
      },
      "CampaignCreate": {
        "type": "object",
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 100},
          "mode": {"type": "string", "enum": ["independent", "threaded"]},
          "target_source": {"type": "string", "enum": ["manual_url", "publish_result"]},
          "target_reference": {"type": "string", "minLength": 1, "maxLength": 2000},
          "template_id": {"type": "string", "minLength": 1, "maxLength": 120},
          "template_revision": {"type": ["integer", "null"], "minimum": 1},
          "profile_refs": {"type": "array", "minItems": 1, "maxItems": 300, "items": {"type": "string", "minLength": 1, "maxLength": 80}, "uniqueItems": true},
          "batch_size": {"type": "integer", "minimum": 1, "maximum": 8, "default": 3},
          "allocation_seed": {"type": "string", "maxLength": 256},
          "start_mode": {"type": "string", "enum": ["manual", "scheduled"], "default": "manual"},
          "scheduled_at": {"type": ["string", "null"], "maxLength": 40}
        },
        "required": ["name", "mode", "target_source", "target_reference", "template_id", "profile_refs"],
        "additionalProperties": false
      },
...
      "CommentCampaignEnvelope": {
        "type": "object", "properties": {"data": {"$ref": "#/components/schemas/CommentCampaignRead"}},
        "required": ["data"], "additionalProperties": false
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
      "GenericSuccess": {
        "description": "当前成功响应；Legacy路由形状可能不统一。",
        "content": {
          "application/json": {
            "schema": {}
          }
        }
      },
      "BadRequest": {
        "description": "请求无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限或本机边界拒绝",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "revision或状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Validation": {
        "description": "业务校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Internal": {
        "description": "内部错误，消息固定且脱敏",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}